Improved method for downloading videos
The old method I was using from some random GitHub Gist was monitoring page traffic for a m3u8 playlist and then grabbing the video from that. This approach had a lot of problems. Fortunately, for every AP Daily Video there is a .json file out there that has the storage links in multiple qualities for the video, and all you need to get it is the hashed ID of the video (which is passed as a URL parameter when accessing AP Classroom videos). So my new method just uses that .json file to grab videos instead of some convoluted traffic monitoring method